Step 4 - Hero Combat
Created Enemy Character (with Weapon) as a target object for setting up combat. Set up collision toggle for Hero Weapon in all attack montages. Added notifications to show collision damage when a collision is successful. Added Hit React and Camera Shake to make the battle experience more immersive and added gameplay cues to trigger montages and visual effects for when the Enemy dies. Set up broadcasting of Hero and Enemy health/rage values to display correctly on screen widgets. Finally, I tweaked the Overlay widget to make it more realistic for gameplay. This is an overly simplified explanation of the ton code required to bring hero combat to life! The following commit will focus on Enemy Artificial Intelligence, so we are not just hitting Enemy dummies that just stand there and get killed.