Step 5: Enemy AI
I created an Enemy Behavior Tree for Gruntling Guardian (the only enemy we have so far) to handle overall logic such as attacking, strafing, and dying. I then added some custom nodes in the Guardian BT to take advantage of motion warping so the enemy mostly faces the hero in the game, especially while attacking. I then set up an enemy damage system so that the enemies can cause damage to our Hero when a successful collision is made. Finally, I added sounds for the enemy dying, weapon swinging, attack cries, and hit sounds for the enemy weapon-to-hero collision. Again, this is an oversimplification of all the code that was needed to bring the Enemy alive. The next commit will focus on additional Hero Combat Abilities.