Step 7 - Ranged Enemy
This commit focused on creating a ranged enemy known as the Gruntling Glacier. Created a behavior tree that tells the Enemy to start an unblockable melee attack when the Hero gets too close and then a ranged attack (fireball) when there is ample distance. Created a projectile class to handle the fireball, along with the damage it deals, the sounds and visual effects. Most of the base info for this Enemy was copied from the Guardian enemy so there was very little coding needed for this section.